# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Booker API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/event": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "List of all events",
                "operationId": "872f3a2ea23a0ec86f5181f74ec6d1b4",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "starttime": {
                                                        "type": "integer"
                                                    },
                                                    "endtime": {
                                                        "type": "integer"
                                                    },
                                                    "boardroom_id": {
                                                        "type": "integer"
                                                    },
                                                    "user_id": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "id": 1,
                                            "description": "Quibusdam sed rerum consectetur consequuntur nulla.",
                                            "starttime": 34930246,
                                            "endtime": 34932046,
                                            "boardroom_id": 1,
                                            "user_id": 1
                                        },
                                        {
                                            "id": 2,
                                            "description": "Quibusdam sed rerum consectetur consequuntur nulla.",
                                            "starttime": 34940246,
                                            "endtime": 34942146,
                                            "boardroom_id": 1,
                                            "user_id": 1
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            },
            "post": {
                "tags": [
                    "Event"
                ],
                "summary": "Creating a new event.",
                "operationId": "da9d2e6694bd97f196ce4da678f68bc2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "description": {
                                                "type": "string"
                                            },
                                            "starttime": {
                                                "type": "integer"
                                            },
                                            "endtime": {
                                                "type": "integer"
                                            },
                                            "boardroom_id": {
                                                "type": "integer"
                                            },
                                            "user_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            },
                            "example": {
                                "description": "New meeting!",
                                "starttime": 34934035,
                                "endtime": 34934036,
                                "boardroom_id": 1,
                                "user_id": 1
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "New description 33"
                                                },
                                                "starttime": {
                                                    "type": "integer",
                                                    "example": 34934035
                                                },
                                                "endtime": {
                                                    "type": "integer",
                                                    "example": 34934036
                                                },
                                                "boardroom_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            }
        },
        "/api/v1/event/{event}": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "Get an event",
                "operationId": "45a2b86483f45b0c880de868804b8693",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Id of the event",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "New description 33"
                                                },
                                                "starttime": {
                                                    "type": "integer",
                                                    "example": 34934035
                                                },
                                                "endtime": {
                                                    "type": "integer",
                                                    "example": 34934036
                                                },
                                                "boardroom_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            },
            "put": {
                "tags": [
                    "Event"
                ],
                "summary": "Update the event",
                "operationId": "13dc252b1f042dcfe56f3f9250a6aa84",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Id of the event",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "description": {
                                                "type": "string"
                                            },
                                            "starttime": {
                                                "type": "integer"
                                            },
                                            "endtime": {
                                                "type": "integer"
                                            },
                                            "boardroom_id": {
                                                "type": "integer"
                                            },
                                            "user_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            },
                            "example": {
                                "description": "New meeting! (updated)",
                                "starttime": 34934035,
                                "endtime": 34934036,
                                "boardroom_id": 1,
                                "user_id": 1
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "New meeting! (updated)"
                                                },
                                                "starttime": {
                                                    "type": "integer",
                                                    "example": 34934035
                                                },
                                                "endtime": {
                                                    "type": "integer",
                                                    "example": 34934036
                                                },
                                                "boardroom_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Event"
                ],
                "summary": "Delete the event",
                "operationId": "2413e0c56943e039190aa86508d051bf",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Id of the event",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}